# docs/api.yaml
basePath: /
host: localhost:3344
info:
  description: Backend of repo https://github.com/vidorg/vid_vue
  license:
    name: MIT
    url: https://github.com/vidorg/vid_backend/blob/master/LICENSE
  termsOfService: https://github.com/vidorg
  title: vid backend
  version: '1.1'
paths:
...
  /v1/user/:
    delete:
      consumes:
      - multipart/form-data
      description: 删除用户所有信息
      operationId: -v1-user-delete
      parameters:
      - *id001
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            \n}\n```"
        '401': *id002
        '404':
          description: '"user not found"'
        '500':
          description: '"user delete failed"'
      security:
      - basicAuth: ''
      summary: 删除用户
      tags:
      - User
    put:
      consumes:
      - multipart/form-data
      description: 更新用户信息
      operationId: -v1-user-put
      parameters:
      - *id001
      - description: '用户名

          *Minimum length* : 8

          *Maximum length* : 30'
        in: formData
        maxLength: 30
        minLength: 8
        name: username
        required: true
        type: string
      - description: 用户性别
        enum:
        - male
        - female
        - unknown
        in: formData
        name: sex
        required: true
        type: string
      - description: '用户简介

          *Minimum length* : 0

          *Maximum length* : 255'
        in: formData
        maxLength: 255
        minLength: 0
        name: profile
        required: true
        type: string
      - description: 用户生日，固定格式为2000-01-01
        in: formData
        name: birth_time
        required: true
        type: string
      - description: 用户手机号码
        in: formData
        name: phone_number
        required: true
        type: string
      - description: 用户头像，默认不修改
        in: formData
        name: avatar
        required: false
        type: file
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            ,\n    \"data\": {\n        \"uid\": 10,\n        \"username\": \"aoihosizora\"\
            ,\n        \"sex\": \"male\",\n        \"profile\": \"Demo profile\",\n\
            \        \"avatar_url\": \"http://localhost:3344/raw/image/default/avatar.jpg\"\
            ,\n        \"birth_time\": \"2001-02-03\",\n        \"authority\": \"\
            normal\",\n        \"phone_number\": \"13512345678\"\n    }\n}\n```"
        '400':
          description: '"request param error" / "request format error" / "request
            body too large" / "username has been used" / "image type not supported"'
        '401': *id002
        '404':
          description: '"user not found"'
        '500':
          description: '"image save failed" / "user update failed"'
      security:
      - basicAuth: ''
      summary: 更新用户
      tags:
      - User
  /v1/user/subscribing:
    delete:
      consumes:
      - multipart/form-data
      description: 取消关注某一用户
      operationId: -v1-user-subscribing-delete
      parameters:
      - *id001
      - description: 取消关注用户id
        in: formData
        name: to
        required: true
        type: integer
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            ,\n    \"data\": {\n        \"me_uid\": 10,\n        \"to_uid\": 3,\n\
            \        \"action\": \"unsubscribe\"\n    }\n}\n```"
        '400':
          description: '"request param error" / "request format error"'
        '401': *id002
        '404':
          description: '"user not found"'
        '500':
          description: '"unsubscribe failed"'
      security:
      - basicAuth: ''
      summary: 取消关注用户
      tags:
      - Subscribe
    put:
      consumes:
      - multipart/form-data
      description: 关注某一用户
      operationId: -v1-user-subscribing-put
      parameters:
      - *id001
      - description: 关注用户id
        in: formData
        name: to
        required: true
        type: integer
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            ,\n    \"data\": {\n        \"me_uid\": 10,\n        \"to_uid\": 3,\n\
            \        \"action\": \"subscribe\"\n    }\n}\n```"
        '400':
          description: '"request param error" / "request format error" / "subscribe
            oneself invalid"'
        '401': *id002
        '404':
          description: '"user not found"'
        '500':
          description: '"subscribe failed"'
      security:
      - basicAuth: ''
      summary: 关注用户
      tags:
      - Subscribe
  /v1/user/{uid}:
    get:
      consumes:
      - multipart/form-data
      description: 查询用户信息
      operationId: -v1-user-{uid}-get
      parameters:
      - description: 用户id
        in: path
        name: uid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            ,\n    \"data\": {\n        \"user\": {\n            \"uid\": 10,\n  \
            \          \"username\": \"aoihosizora\",\n            \"sex\": \"male\"\
            ,\n            \"profile\": \"Demo profile\",\n            \"avatar_url\"\
            : \"http://localhost:3344/raw/image/default/avatar.jpg\",\n          \
            \  \"birth_time\": \"2001-02-03\",\n            \"authority\": \"normal\"\
            ,\n            \"phone_number\": \"13512345678\"\n        },\n       \
            \ \"extra\": {\n            \"subscribing_cnt\": 3,\n            \"subscriber_cnt\"\
            : 2,\n            \"video_cnt\": 3\n        }\n    }\n}\n```"
        '400':
          description: '"request param error"'
        '404':
          description: '"user not found"'
      summary: 查询用户
      tags:
      - User
  /v1/user/{uid}/subscriber:
    get:
      consumes:
      - multipart/form-data
      description: 查询用户所有粉丝，返回分页数据
      operationId: -v1-user-{uid}-subscriber-get
      parameters:
      - description: 查询的用户id
        in: path
        name: uid
        required: true
        type: integer
      - description: 分页
        in: query
        name: page
        required: false
        type: integer
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            ,\n    \"data\": {\n        \"count\": 1,\n        \"page\": 1,\n    \
            \    \"data\": [\n            {\n                \"uid\": 10,\n      \
            \          \"username\": \"aoihosizora\",\n                \"sex\": \"\
            male\",\n                \"profile\": \"Demo profile\",\n            \
            \    \"avatar_url\": \"http://localhost:3344/raw/image/default/avatar.jpg\"\
            ,\n                \"birth_time\": \"2001-02-03\",\n                \"\
            authority\": \"normal\",\n                \"phone_number\": \"13512345678\"\
            \n            }\n        ]\n    }\n}\n```"
        '400':
          description: '"request param error"'
        '404':
          description: '"user not found"'
      summary: 用户粉丝
      tags:
      - Subscribe
  /v1/user/{uid}/subscribing:
    get:
      consumes:
      - multipart/form-data
      description: 查询用户所有关注，返回分页数据
      operationId: -v1-user-{uid}-subscribing-get
      parameters:
      - description: 查询的用户id
        in: path
        name: uid
        required: true
        type: integer
      - description: 分页
        in: query
        name: page
        required: false
        type: integer
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            ,\n    \"data\": {\n        \"count\": 1,\n        \"page\": 1,\n    \
            \    \"data\": [\n            {\n                \"uid\": 10,\n      \
            \          \"username\": \"aoihosizora\",\n                \"sex\": \"\
            male\",\n                \"profile\": \"Demo profile\",\n            \
            \    \"avatar_url\": \"http://localhost:3344/raw/image/default/avatar.jpg\"\
            ,\n                \"birth_time\": \"2001-02-03\",\n                \"\
            authority\": \"normal\",\n                \"phone_number\": \"13512345678\"\
            \n            }\n        ]\n    }\n}\n```"
        '400':
          description: '"request param error"'
        '404':
          description: '"user not found"'
      summary: 用户关注
      tags:
      - Subscribe
...
  /v1/user?page:
    get:
      consumes:
      - multipart/form-data
      description: 管理员查询所有用户，返回分页数据，Admin
      operationId: -v1-user?page-get
      parameters:
      - *id001
      - description: 分页
        in: query
        name: page
        required: false
        type: integer
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            ,\n    \"data\": {\n        \"count\": 1,\n        \"page\": 1,\n    \
            \    \"data\": [\n            {\n                \"uid\": 10,\n      \
            \          \"username\": \"aoihosizora\",\n                \"sex\": \"\
            male\",\n                \"profile\": \"Demo profile\",\n            \
            \    \"avatar_url\": \"http://localhost:3344/raw/image/default/avatar.jpg\"\
            ,\n                \"birth_time\": \"2001-02-03\",\n                \"\
            authority\": \"normal\",\n                \"phone_number\": \"13512345678\"\
            \n            }\n        ]\n    }\n}\n```"
        '400':
          description: '"request param error"'
        '401': *id002
      security:
      - basicAuth: ''
      summary: 查询所有用户
      tags:
      - User
      - Administration
...
swagger: '2.0'